refactor: MemberService 리팩토링
* modify 메서드에 flush() 추가
* delete 메서드에 비밀번호 비교 로직 추가

+++ backend/src/main/java/com/project/backend/domain/member/service/MemberService.java
@@ -97,9 +97,13 @@ public class MemberService {
         member.setGender(gender);
         member.setNickname(nickname);
         member.setBirth(birth);
+        memberRepository.flush();
     }
 
-    public void delete(Member member) {
+    public void delete(Member member, String password) {
+        if(!passwordEncoder.matches(password, member.getPassword())) {
+            throw new MemberException(INCORRECT_PASSWORD);
+        }
         memberRepository.delete(member);
     }
